🚧 PX8PZT task: apply external agent result
