+++ .agentplane/tasks/202609031717-PX8PZT/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202609031717-PX8PZT/port-the-minimal-missing-clean-core-lifecycle-bo",
   "created_at": "2026-09-03T17:26:07.913Z",
-  "diffstat_sha256": "sha256:1be280076972d2c10235f277d940763202341a52a0af204e8f6918046b6ac6f2",
+  "diffstat_sha256": "sha256:dd73649a0ee814da241fa0de48ef14830a22d66931b5bc712110745f387164d7",
   "last_verified_at": null,
   "provider": {
     "hostname": "github.com",
