+++ packages/agentplane/src/commands/task/direct-task-verification.test.ts
@@ -26,6 +26,7 @@ import {
   runDirectTaskVerification,
 } from "./direct-task-verification.js";
 import { resolveEvidenceOnlyReworkCommit } from "./evidence-only-rework-commit.js";
+import "./direct-task-verification.sequence.cases.js";
 
 const TASK_ID = "202607290000-RF10A1";
 const roots: string[] = [];
@@ -366,121 +367,6 @@ describe("direct task verification", () => {
     });
     expect(invocation?.env).not.toHaveProperty("AGENTPLANE_AGENT_MODE");
     expect(invocation?.env).not.toHaveProperty("AGENTPLANE_RUNTIME_ACTIVE_BIN");
-  });
-
-  it("runs a safe declared check sequence in order without a shell", async () => {
-    const cwd = await root();
-    mocks.runProcess
-      .mockResolvedValueOnce({ exitCode: 0, stdout: "generated", stderr: "" })
-      .mockResolvedValueOnce({ exitCode: 0, stdout: "fresh", stderr: "" });
-    const check = "bun run docs:readme-header:generate && bun run docs:readme-header:check";
-
-    const result = await runDirectTaskVerification({
-      command: command(cwd),
-      task: { verify: [check], task_kind: "code", mutation_scope: "code" },
-      task_id: TASK_ID,
-      cwd,
-      run_process: mocks.runProcess,
-    });
-
-    expect(result).toMatchObject({
-      status: "passed",
-      reason: null,
-      checks: [{ command: check, script: null, exit_code: 0, stdout_tail: "generated\nfresh" }],
-    });
-    expect(mocks.runProcess).toHaveBeenNthCalledWith(
-      1,
-      expect.objectContaining({
-        command: "bun",
-        args: ["run", "docs:readme-header:generate"],
-      }),
-    );
-    expect(mocks.runProcess).toHaveBeenNthCalledWith(
-      2,
-      expect.objectContaining({ command: "bun", args: ["run", "docs:readme-header:check"] }),
-    );
-  });
-
-  it("stops a declared check sequence on first failure and shares its timeout budget", async () => {
-    const cwd = await root();
-    mocks.runProcess
-      .mockResolvedValueOnce({ exitCode: 0, stdout: "first", stderr: "" })
-      .mockResolvedValueOnce({ exitCode: 7, stdout: "", stderr: "second failed" });
-    const nowValues = [1000, 1000, 1250, 1250];
-    const check = "bun run first && bun run second && bun run should-not-run";
-
-    const result = await runDirectTaskVerification({
-      command: command(cwd),
-      task: { verify: [], task_kind: "code", mutation_scope: "code" },
-      task_id: TASK_ID,
-      cwd,
-      additional_commands: [{ command: check, timeout_ms: 1000 }],
-      additional_only: true,
-      run_process: mocks.runProcess,
-      now: () => nowValues.shift() ?? 1250,
-    });
-
-    expect(result).toMatchObject({
-      status: "failed",
-      reason: `Declared check failed: ${check}`,
-      checks: [{ command: check, exit_code: 7 }],
-    });
-    expect(mocks.runProcess).toHaveBeenCalledTimes(2);
-    expect(mocks.runProcess).toHaveBeenNthCalledWith(
-      1,
-      expect.objectContaining({ timeoutMs: 1000 }),
-    );
-    expect(mocks.runProcess).toHaveBeenNthCalledWith(
-      2,
-      expect.objectContaining({ timeoutMs: 750 }),
-    );
-  });
-
-  it("rejects malformed or unsafe declared check sequences before execution", async () => {
-    for (const check of [
-      "bun run first &&",
-      "&& bun run second",
-      "bun run first && && bun run second",
-      "bun run first&&bun run second",
-      "bun run first || bun run second",
-      "bun run first; bun run second",
-      "bun run first & bun run second",
-    ]) {
-      const cwd = await root();
-      const result = await runDirectTaskVerification({
-        command: command(cwd),
-        task: { verify: [check], task_kind: "code", mutation_scope: "code" },
-        task_id: TASK_ID,
-        cwd,
-        run_process: mocks.runProcess,
-      });
-
-      expect(result).toMatchObject({
-        status: "unsupported",
-        reason: `Unsupported declared check: ${check}`,
-      });
-    }
-    expect(mocks.runProcess).not.toHaveBeenCalled();
-  });
-
-  it("keeps quoted ampersands inside one structured argument", async () => {
-    const cwd = await root();
-    mocks.runProcess.mockResolvedValueOnce({ exitCode: 0, stdout: "1 pass", stderr: "" });
-    const check = "bun test 'a && b'";
-
-    const result = await runDirectTaskVerification({
-      command: command(cwd),
-      task: { verify: [check], task_kind: "code", mutation_scope: "code" },
-      task_id: TASK_ID,
-      cwd,
-      run_process: mocks.runProcess,
-    });
-
-    expect(result.status).toBe("passed");
-    expect(mocks.runProcess).toHaveBeenCalledOnce();
-    expect(mocks.runProcess).toHaveBeenCalledWith(
-      expect.objectContaining({ command: "bun", args: ["test", "a && b"] }),
-    );
   });
 
   it("rejects successful Bun processes that report zero executed tests", async () => {
